pyproject.toml
[project]
name = "ats-checker"
version = "0.1.0"
description = "ATS Resume Analyzer - Score and optimize your resume for job descriptions using AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scikit-learn>=1.6.1",
    "streamlit>=1.45.1",
    "pdfplumber>=0.11.0",
    "ollama>=0.4.0",
    "python-docx>=1.1.0",
    "weasyprint>=62.0",
    "Markdown>=3.5.0",
    "pydantic>=2.9.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
ats-checker = "ats_checker.cli:main"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=ats_checker --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ats_checker"]

[tool.uv]
index-url = "https://pypi.org/simple"
